moved tests to its own directory
